feat(linux): add an additional condition for Amazon Linux
This commit modifies the setup-Linux.yml file to include an additional condition for Amazon Linux. The new condition checks if the ansible_distribution_major_version is equal to "2". This change ensures that the specified command is executed only when both ansible_distribution is "Amazon" and ansible_distribution_major_version is "2".